Wanda.Executions.Server (Wanda v1.2.0)

Represents the execution of the CheckSelection(s) on the target nodes/agents of a cluster Orchestrates facts gathering on the targets - issuing execution and receiving back facts - and following check evaluations.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_execution(execution_id, group_id, targets, target_type, env, config \\ [])

Starts a check execution.

The checks are filtered leveraging the env and evaluating the when condition using a best-effort approach. If non-existing check IDs are provided inside a target, they will get filtered away.

start_link(opts)