WandaWeb.Router.Helpers (Wanda v0.1.0)

Module with named helpers generated from WandaWeb.Router.

Functions

catalog_path(conn_or_endpoint, action)

catalog_path(conn_or_endpoint, action, params)

catalog_url(conn_or_endpoint, action)

catalog_url(conn_or_endpoint, action, params)

execution_path(conn_or_endpoint, action)

execution_path(conn_or_endpoint, action, params)

execution_path(conn_or_endpoint, action, id, params)

execution_url(conn_or_endpoint, action)

execution_url(conn_or_endpoint, action, params)

execution_url(conn_or_endpoint, action, id, params)

path(data, path)

Generates the path information including any necessary prefix.

render_spec_path(conn_or_endpoint, action)

render_spec_path(conn_or_endpoint, action, params)

render_spec_url(conn_or_endpoint, action)

render_spec_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

swagger_ui_path(conn_or_endpoint, action)

swagger_ui_path(conn_or_endpoint, action, params)

swagger_ui_url(conn_or_endpoint, action)

swagger_ui_url(conn_or_endpoint, action, params)

url(data)

Generates the connection/endpoint base URL without any path information.