Wanda.Executions.Gathering (wanda v0.1.0)

Facts gathering functional core.

Functions

all_agents_sent_facts?(agents_gathered, targets)

@spec all_agents_sent_facts?([String.t()], [Wanda.Executions.Target.t()]) :: boolean()

Check if all the agents have sent the facts

put_gathered_facts(gathered_facts, agent_id, facts)

@spec put_gathered_facts(map(), String.t(), [
  Wanda.Executions.Fact.t() | Wanda.Executions.FactError.t()
]) ::
  map()

Adds gathered facts of an agent.

put_gathering_timeouts(gathered_facts, timed_out_agents)

@spec put_gathering_timeouts(map(), [Wanda.Executions.Target.t()]) :: map()

Adds timeout data to gathered facts.

target?(targets, agent_id)

@spec target?([Wanda.Executions.Target.t()], String.t()) :: boolean()

Check if an agent is a target of an execution